Split bridge tests into own module, document networkx ordering pitfall
- Move the wmg-to-neural-lam bridge tests (validator-backed spec tests,
  to_torch_tensors_on_disk behaviour tests and the _graph_to_edge_tensors
  unit tests) into tests/test_save_torch_tensors.py. tests/test_save.py now
  only holds the tests for the deprecated to_pyg path, so it can be deleted
  together with that functionality.
- Add a test, using no weather-model-graphs functionality, documenting that
  pyg's from_networkx numbers nodes by networkx insertion order rather than
  by node label, so inserting nodes out of label order yields edge-index
  values that do not correspond to the labels. The docstring records the two
  ways weather-model-graphs guards against this (sort_nodes_in_graph before
  from_networkx in the deprecated path; explicit label-sorted maps in
  _graph_to_edge_tensors).
- Reword the edge-sort comment in _graph_to_edge_tensors to separate the two
  ordering concerns: node index values (safe here by construction via the
  sorted-label maps) and edge column order (fixed by the sort for
  byte-for-byte reproducible output).
